Refactor paymentReceivedSheet using sheet component and cleanup with cardstack components
