+++ cardstack/src/screens/index.ts
@@ -14,3 +14,4 @@ export { default as LoadingOverlayScreen } from './LoadingOverlayScreen';
 export { default as WelcomeScreen } from './WelcomeScreen/WelcomeScreen';
 export { default as ImportSeedSheet } from './sheets/ImportSeed/ImportSeedSheet';
 export { default as CollectibleSheet } from './sheets/Collectible/CollectibleSheet';
+export { default as PaymentReceivedSheet } from './sheets/PaymentReceived/PaymentReceivedSheet';
