--- src/components/expanded-state/payment-item-details/index.ts
@@ -1,1 +1,0 @@
-export * from './PaymentItemDetails';
