+++ src/screens/ExpandedAssetSheet.js
@@ -10,7 +10,6 @@ import {
   ChartExpandedState,
   LifetimeEarningsExpandedState,
   LiquidityPoolExpandedState,
-  MerchantSpendReceivedExpandedState,
   MerchantTransactionExpandedState,
   PaymentConfirmationExpandedState,
   SupportAndFeedsState,
@@ -32,7 +31,6 @@ const ScreenTypes = {
   [ExpandedMerchantRoutes.paymentRequest]: PaymentRequestExpandedState,
   merchantTransaction: MerchantTransactionExpandedState,
   paymentConfirmationTransaction: PaymentConfirmationExpandedState,
-  merchantSpendReceivedExpandedState: MerchantSpendReceivedExpandedState,
   supportAndFees: SupportAndFeedsState,
 };
